+++ mobile/lib/deep-links.ts
@@ -1,0 +1,7 @@
+import { SITE_URL } from "./api"
+
+export const WEEK_VIEW_WEB_URL = `${SITE_URL}/?week=1`
+
+export const CUENTA_WEB_URL = `${SITE_URL}/cuenta`
+
+export const CUENTA_EXPORT_HINT = `${SITE_URL}/cuenta`